+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 
 * Dialog now takes a prop of diableBackground which is true by default
 
+## Bug Fixes
+
+* Fixes alignment issue with SplitButton when using anchors.
+
 # 0.7.1
 
 ## Updates