+++ CHANGELOG.md
@@ -10,6 +10,10 @@
   * `border` - allows developers to disable border.
   * `padding` - allows developers to have control over padding size.
 
+## Improve Dialog
+
+* Dialog now takes a prop of diableBackground which is true by default
+
 ## Bug Fixes
 
 * Fixes alignment issue with SplitButton when using anchors.