+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## New Components
 
+* Portrait
 * Tile - an alternative to Pod, with a drop shadow.
 
 ## Layout Updates
@@ -9,6 +10,10 @@
 * Pod component now receives two additional props:
   * `border` - allows developers to disable border.
   * `padding` - allows developers to have control over padding size.
+
+## Improve Dialog
+
+* Dialog now takes a prop of diableBackground which is true by default
 
 ## Bug Fixes